Sahar intra

+++ lib/src/data/remote/api/api_request_dtos/merchant_flow/post_bid_request_dto.dart
@@ -1,7 +1,7 @@
 class PostBidRequestDto {
   final int orderId;
   final int merchantId;
-  final double bidAmount;
+  final int bidAmount;
   final int createdBy;
 
   PostBidRequestDto({
